[python] Move miscellaneous wrappers to separate files
